+++ client/src/App.tsx
@@ -6,7 +6,6 @@ import About from "./views/About";
 import NotFound from "./views/NotFound";
 import Charts from "./views/Charts";
 import Terms from "./views/Terms";
-import Learn from "./views/Learn";
 import Game from "./views/Game";
 import Customize from "./views/Customize";
 import * as ROUTES from "./data/routes";
@@ -20,7 +19,6 @@ const App = () => {
         <Route path={ROUTES.ABOUT_US} element={<About />}/>
         <Route path={ROUTES.CHARTS} element={<Charts />}/>
         <Route path={ROUTES.TERMS} element={<Terms />}/>
-        <Route path={ROUTES.LEARN} element={<Learn />}/>
         <Route path={ROUTES.GAME} element={<Game />}/>
         <Route path={ROUTES.CUSTOMIZE} element={<Customize />}/>
         <Route path={ROUTES.NOT_FOUND} element={<NotFound />}/>
