+++ client/src/components/Card/Card.module.css
@@ -1,7 +1,7 @@
 .card {
-    width: 90px;
-    height: 195px;
-    margin: 10px;
+    width: 130px;
+    height: 250px;
+    margin: 20px;
     padding: 0.5em 1.5em;
     background: rgb(230, 230, 230);
     border-radius: 15px;
@@ -10,9 +10,9 @@
   }
   
   .hiddenCard {
-    width: 85px;
-    height: 185px;
-    margin: 10px;
+    width: 125px;
+    height: 240px;
+    margin: 20px;
     padding: 0.5em 1.5em;
     background-image: linear-gradient(176deg, #ffffff 8.33%, #ff0000 8.33%, #ff0000 50%, #ffffff 50%, #ffffff 58.33%, #ff0000 58.33%, #ff0000 100%);
     background-size: 60.00px 4.20px;
@@ -32,7 +32,7 @@
   
   .value {
     font-size: 300%;
-    margin: 0;
+    margin-top: 2px;
   }
   
   .suit {
