+++ client/src/components/Card/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './styles/Card.module.css';
+import styles from './Card.module.css';
 
 type CardProps = {
   value: string;
