--- client/src/components/Cards/index.tsx
@@ -1,7 +1,0 @@
-const Cards = () => {
-    return (
-        <div></div>
-    )
-}
-
-export default Cards;
