+++ client/src/components/Controls/index.tsx
@@ -1,0 +1,46 @@
+import React, { useState } from 'react';
+import styles from './Controls.module.css';
+
+type ControlsProps = {
+  gameState: number,
+  buttonState: any,
+  playEvent: any,
+  hitEvent: any,
+  standEvent: any,
+  resetEvent: any
+};
+
+const Controls: React.FC<ControlsProps> = ({ gameState, buttonState, playEvent, hitEvent, standEvent, resetEvent }) => {
+  const [inputStyle, setInputStyle] = useState(styles.input);
+
+  const onBetClick = () => {
+    playEvent();
+  }
+
+  const getControls = () => {
+    if (gameState === 0) {
+      return (
+        <div className={styles.controlsContainer}>
+          <button onClick={() => onBetClick()} className={styles.button}>Play</button>
+        </div>
+      );
+    }
+    else {
+      return (
+        <div className={styles.controlsContainer}>
+          <button onClick={() => hitEvent()} disabled={buttonState.hitDisabled} className={styles.button}>Hit</button>
+          <button onClick={() => standEvent()} disabled={buttonState.standDisabled} className={styles.button}>Stand</button>
+          <button onClick={() => resetEvent()} disabled={buttonState.resetDisabled} className={styles.button}>Reset</button>
+        </div>
+      );
+    }
+  }
+
+  return (
+    <>
+      {getControls()}
+    </>
+  );
+}
+
+export default Controls;
