+++ client/src/components/Hand/Hand.module.css
@@ -4,12 +4,17 @@
     align-items: center;
     flex-direction: column;
     margin: 0.5em;
+    min-height: 0;
   }
   
   .cardContainer {
     display: flex;
     justify-content: center;
     flex-wrap: wrap;
+    background-color: black;
+    border: 5px solid white;
+    border-radius: 15px;
+    box-shadow: 0px 1px 10px rgb(0, 0, 0);
   }
   
   .title {
