+++ client/src/components/Title/index.tsx
@@ -1,21 +1,17 @@
 import { useNavigate } from "react-router-dom";
 import "./title.css";
-import { LEARN, GAME, CUSTOMIZE } from "../../data/routes";
+import { GAME, CUSTOMIZE } from "../../data/routes";
 
 const Title = () => {
   const navigate = useNavigate();
 
-  const handleCalculateClick = () => {
-    navigate(LEARN);
-  };
-
   const handlePlayGameClick = () => {
     navigate(GAME);
-  }
+  };
 
   const handleCustomizeClick = () => {
     navigate(CUSTOMIZE);
-  }
+  };
 
   return (
     <div className="background">
@@ -25,15 +21,10 @@ const Title = () => {
           Ever wanted to learn how to become a true professional at Blackjack?
           Well click 'Learn the Game' below to learn how you can!
         </p>
-        <div className='column'>
-          <div className='row'>
-            <button className="btn-hover" onClick={handleCalculateClick}>
-              Learn the Game
-            </button>
-            <button className="btn-hover" onClick={handlePlayGameClick}>
-              Play the Game
-            </button>
-          </div>
+        <div className="row">
+          <button className="btn-hover" onClick={handlePlayGameClick}>
+            Play the Game
+          </button>
           <button className="btn-hover" onClick={handleCustomizeClick}>
             Customize Table
           </button>
