+++ client/src/data/routes.tsx
@@ -2,7 +2,6 @@ export const HOME = "/blackjack_probability";
 export const ABOUT_US = "/about-us";
 export const CHARTS = "/charts";
 export const TERMS = "/terms";
-export const LEARN = "/learn";
 export const GAME = "/game";
 export const CUSTOMIZE = "/customize";
 export const NOT_FOUND = "*";
