--- client/src/views/Game/styles/Game.module.css
@@ -1,4 +1,0 @@
-.gameBackground {
-    background-color: green;
-    height: 750px;
-}
