--- client/src/views/Learn/index.tsx
@@ -1,9 +1,0 @@
-const Learn = () => {
-    return (
-        <div>
-            Learn Mode
-        </div>
-    )
-}
-
-export default Learn;
